Frame index start at 1

+++ src/test/test_sphere.cc
@@ -18,13 +18,13 @@ int n_points = 12;
 
 //------------------------------------------------------------------------------
 
-void batch(unsigned nbp, unsigned repeat)
+void batch(unsigned long nbp, int repeat)
 {
-    unsigned iterations = 0;
+    size_t iterations = 0;
     real energy = INFINITY;
     real distance = 0;
     
-    printf("%4i pts :", nbp);
+    printf("%4lu pts :", nbp);
     printf(" %6.4f :", S.expectedDistance(nbp));
     
     for ( int m=0; m < repeat; ++m )
@@ -40,7 +40,7 @@ void batch(unsigned nbp, unsigned repeat)
     
     printf("   distance %9.6f",    distance);
     printf("   energy %14.5f",     energy);
-    printf("   iterations %12u\n", iterations);
+    printf("   iterations %12lu\n", iterations);
 }
 
 //------------------------------------------------------------------------------
@@ -173,12 +173,10 @@ int main(int argc, char* argv[])
     
     if ( argc == 3 ) 
     {
-        unsigned min = (unsigned)strtoul(argv[1], 0, 10);
-        unsigned max = (unsigned)strtoul(argv[2], 0, 10);
-        
-        for ( size_t nbp = min; nbp < max; nbp += 7)
+        unsigned long min = strtoul(argv[1], 0, 10);
+        unsigned long max = strtoul(argv[2], 0, 10);
+        for ( unsigned long nbp = min; nbp < max; nbp += 7)
             batch(nbp, 16);
-        
         return EXIT_SUCCESS;
     }
